+++ docs/conf.py
@@ -21,9 +21,9 @@ author = "Siarhei Skuratovich"
 
 # The version info
 # The short X.Y version
-version = "0.5.0"
+version = "0.6.0"
 # The full version, including alpha/beta/rc tags
-release = "0.5.0"
+release = "0.6.0"
 
 # -- General configuration ---------------------------------------------------